Equation Generator
Added an algorithm to randomly generate an expression and gather an answer based off the operations in the expression.

The expression will be outputted into the GUI, and then a check is made when the user enters in a value in the textbox. If the value equals the correct answer, it will tell the user they are correct, and if not, they will get summoned to the gazebo.